DeplAI_old/requirements.txt
# Core Frameworks
# For building applications with LLMs
langchain==0.1.20
# Community integrations for LangChain
langchain-community==0.0.38
# OpenAI integration for LangChain
langchain-openai==0.1.6
# A data framework for LLM applications to connect custom data sources to LLMs
llama-index>=0.8.0
# For creating web apps for machine learning and data science
streamlit>=1.26.0

# -------------------------------------------------------------

# Vector Database
# An open-source embedding database
chromadb==0.4.24
# ChromaDB vector store for LlamaIndex
llama-index-vector-stores-chroma~=0.2.0

# -------------------------------------------------------------

# LLM & Embeddings
# Python client for the OpenAI API
openai>=1.3.0
# For state-of-the-art sentence, text and image embeddings
sentence-transformers>=2.2.2
# Client library for the Hugging Face Hub
huggingface-hub>=0.16.4
# State-of-the-art machine learning for PyTorch, TensorFlow, and JAX
transformers>=4.33.2

# -------------------------------------------------------------

# Messaging Queue
# Pure-Python implementation of the AMQP 0-9-1 protocol (for RabbitMQ)
pika==1.3.2

# -------------------------------------------------------------

# Data Processing & Validation
# Data validation and settings management using Python type hints
pydantic~=2.8.0
# For the creation, manipulation, and study of complex networks
networkx>=3.1
# An implementation of the JSON Schema specification for Python
jsonschema>=4.19.0

# -------------------------------------------------------------

# Storage & Caching
# Python client for Redis, an in-memory data structure store
redis==5.0.4
# Used for database operations and object-relational mapping
sqlalchemy>=2.0.20

# -------------------------------------------------------------

# Terraform Tools Integration
# A parser for HCL2, used for Terraform files
python-hcl2>=4.3.0
# A modern and designer-friendly templating language for Python
jinja2>=3.1.2
# A generic syntax highlighter
pygments>=2.16.1

# -------------------------------------------------------------

# Visualization
# For rendering graphs in Streamlit applications
streamlit-agraph>=0.0.45
# An interactive graphing library for Python
plotly>=5.16.1
# For drawing cloud system architecture diagrams in Python code
diagrams

# -------------------------------------------------------------

# Utilities
# Reads key-value pairs from a .env file and sets them as environment variables
python-dotenv==1.0.1
# A simple, yet elegant, HTTP library for Python
requests>=2.31.0
# For pulling data out of HTML and XML files
beautifulsoup4>=4.12.2
# The fundamental package for scientific computing with Python
numpy>=1.25.2
# Fast, powerful, flexible and easy to use open source data analysis and manipulation tool
pandas>=2.1.0
# A fast, extensible progress bar for Python and CLI
tqdm>=4.66.1
# Makes ANSI escape character sequences work on Windows
colorama>=0.4.6

# -------------------------------------------------------------

# AWS SDK (for validation)
# The AWS SDK for Python
boto3>=1.28.40

# -------------------------------------------------------------

# Web Scraping with JavaScript Rendering
# selenium>=4.10.0
# webdriver-manager>=4.0.0 

# -------------------------------------------------------------

# YAML parsing
# YAML parser and emitter for Python
pyyaml>=6.0.1

# -------------------------------------------------------------

# Tavily API
# Python client for the Tavily Search API
tavily-python

# -------------------------------------------------------------

# GitHub API
# Python library to access the GitHub API v3
PyGithub

# -------------------------------------------------------------

# New package
# Hugging Face embeddings for LlamaIndex
llama-index-embeddings-huggingface

# -------------------------------------------------------------

# Testing
# A framework for writing software tests
pytest

# -------------------------------------------------------------

#Azure Cost manegement api
# Azure Cost Management client library for Python
azure-mgmt-costmanagement>=0.1.0

#GCP Cost management api
# Google Cloud Billing API client for Python
google-cloud-billing>=1.11.0
